// src/services/accounts/Centurion.Accounts.DashboardSpa/AngularClient/src/core/models/claim-names.model.ts
export enum ClaimNames {
  name = "name",
  email = "email",
  id = "id",
  avatar = "avatar",
  discriminator = "discriminator",
  stripeCustomerId = "stripe_customer_id",
  discordId = "discord_id",
  discordAccessTokenToken = "discord_access_token",
  discordRefreshTokenToken = "discord_refresh_token",
  discordAvatar = "avatar",
  discordRoleName = "discord_role_name",
  discordDiscriminator = "discriminator",
  licenseKey = "license_key",
  licenseKeyExpiry = "key_expiry",
  productVersion = "product_version",
  ownDashboardId = "own_dashboard_id",
  joinedDashboardId = "joined_dashboard_id",
  currentDashboardId = "curr_dashboard_id",
  currentDashboardHostingMode = "curr_dashboard_host_mode",
  currentDashboardDomain = "curr_dashboard_domain",
  permission = "permission",

  roleName = "role_name",
  roleId = "role_id",
}